requirements.txt
streamlit
pillow
numpy
scikit-learn
torch
torchvision
pandas
numpy
joblib
fastai